consumer-app-error/requirements.txt
kafka-python
psycopg2-binary
SQLAlchemy==2.0.36
python-dotenv
